Placed character creation in a separate class and also created a character object and assigned it to a variable
